lots of trivial changes (int -> NSInteger) and such; want to update this from 10.4 to 10.6 universal i386/x86_64 binary

+++ HUDWindow.m
@@ -12,7 +12,7 @@
 
 
 - (id)initWithContentRect:(NSRect)contentRect 
-                styleMask:(unsigned int)styleMask 
+                styleMask:(NSUInteger)styleMask 
                   backing:(NSBackingStoreType)bufferingType 
                     defer:(BOOL)flag 
 {
