+++ PDURLTextFieldCell.m
@@ -116,7 +116,7 @@
 }
 
 - (void)selectWithFrame:(NSRect)aRect inView:(NSView *)controlView editor:(NSText *)textObj 
-		delegate:(id)anObject start:(int)selStart length:(int)selLength
+		delegate:(id)anObject start:(NSInteger)selStart length:(NSInteger)selLength
 {
     [super selectWithFrame:[self textRectForFrame:aRect] inView:controlView editor:textObj 
 			delegate:anObject start:selStart length:selLength];
